requirements.txt
streamlit
request
load_dotenv
io
time
ffmpeg  
ipython
pydub  